Server2/requirements.txt
Flask>=2.0.3
requests>=2.26.0
pandas>=1.3.3
numpy>=1.21.2
scikit-learn>=0.24.2
Flask
flask-cors
pandas
numpy
scikit-learn
PyPDF2
python-dotenv
google-generativeai
